simplify implementation by removing harmony patch for AutoEngineerPlanner
